refactor: add shared scroll function to utils
